mcauth/requirements.txt
httpx>=0.24.0
cryptography>=41.0.0
PyJWT=>2.10.1